+++ api/matlab/arrus/mexcuda/rawReorg.cu
@@ -9,6 +9,11 @@
  *
  * NOTE:
  * - the block of threads must have OEM_N_CHAN x OEM_N_CHAN size (must be square for transposition)
+ * - for dimension consistency the blocksPerGrid size should be {1, N, 1} but for proper support of 
+     large input data, blocksPerGrid is {N, 1, 1}. Max allowed value of blocksPerGrid[0] is much 
+     higher than in case of blocksPerGrid[1]: 2^31-1 vs 2^16-1. For proper interpretation of the 
+     grid dimensions consider blockIdx.x & blockDim.x as if they were blockIdx.y & blockDim.y, 
+     respectively.
  * - is operation on float2 space coalescent?
  */
 
@@ -22,8 +27,8 @@ __global__ void rawReorgCplx( float2 *out,
     size_t idx;
     
     // Input indices
-    int iChanIn = threadIdx.x;                           // gridSize[0] = 1 -> blockIdx.x = 0
-    int iSampIn = blockIdx.y * blockDim.y + threadIdx.y; // real and imag pair as a single sample
+    int iChanIn = threadIdx.x;
+    int iSampIn = blockIdx.x * blockDim.x + threadIdx.y; // real and imag pair as a single sample
     
     // Copy input to shared
     idx = iSampIn*2*OEM_N_CHAN + iChanIn;
@@ -34,7 +39,7 @@ __global__ void rawReorgCplx( float2 *out,
     __syncthreads();
     
     // Output indices
-    int iSampAux = blockIdx.y * blockDim.y + threadIdx.x;
+    int iSampAux = blockIdx.x * blockDim.x + threadIdx.x;
     int iChanAux = threadIdx.y;
     
     int iChunkIn = iSampAux / nSampOut;
@@ -65,7 +70,7 @@ __global__ void rawReorgReal( float *out,
     
     // Input indices
     int iChanIn = threadIdx.x;
-    int iSampIn = blockIdx.y * blockDim.y + threadIdx.y;
+    int iSampIn = blockIdx.x * blockDim.x + threadIdx.y;
     
     // Copy input to shared
     idx = iSampIn*OEM_N_CHAN + iChanIn;
@@ -75,7 +80,7 @@ __global__ void rawReorgReal( float *out,
     __syncthreads();
     
     // Output indices
-    int iSampAux = blockIdx.y * blockDim.y + threadIdx.x;
+    int iSampAux = blockIdx.x * blockDim.x + threadIdx.x;
     int iChanAux = threadIdx.y;
     
     int iChunkIn = iSampAux / nSampOut;
@@ -156,7 +161,7 @@ void mexFunction(int nlhs, mxArray * plhs[],
         out = mxGPUCreateGPUArray(nDimOut, dimOut, mxSINGLE_CLASS, mxCOMPLEX, MX_GPU_INITIALIZE_VALUES);
         float2 * dev_out = static_cast<float2 *>(mxGPUGetData(out));
         
-        dim3 blocksPerGrid = {1, nSampIn / 2 / threadsPerBlock.y, 1};
+        dim3 blocksPerGrid = {nSampIn / 2 / threadsPerBlock.y, 1, 1};
         rawReorgCplx<<<blocksPerGrid, threadsPerBlock>>>( dev_out, dev_in, dev_reorgMap, nSampOut);
     }
     else {
@@ -168,7 +173,7 @@ void mexFunction(int nlhs, mxArray * plhs[],
         out = mxGPUCreateGPUArray(nDimOut, dimOut, mxSINGLE_CLASS, mxREAL, MX_GPU_INITIALIZE_VALUES);
         float * dev_out = static_cast<float *>(mxGPUGetData(out));
         
-        dim3 blocksPerGrid = {1, nSampIn / threadsPerBlock.y, 1};
+        dim3 blocksPerGrid = {nSampIn / threadsPerBlock.y, 1, 1};
         rawReorgReal<<<blocksPerGrid, threadsPerBlock>>>( dev_out, dev_in, dev_reorgMap, nSampOut);
     }