+++ arrus/core/devices/us4r/us4oem/Us4OEMImpl.cpp
@@ -326,7 +326,7 @@ std::pair<size_t, float> Us4OEMImpl::scheduleReceiveDDC(size_t outputAddress,
 size_t Us4OEMImpl::scheduleReceiveRF(size_t outputAddress, uint32 startSample, uint32 endSample, uint16 entryId,
                                      const TxRxParameters &op, uint16 rxMapId) {
     const uint32 startSampleRaw = startSample * op.getRxDecimationFactor();
-    const uint32 sampleRxOffset = ius4oem->GetTxOffset();
+    const uint32 sampleRxOffset = descriptor.getSampleTxStart();
     const size_t nSamples = endSample - startSample;
     const size_t nSamplesRaw = nSamples;
     const size_t sampleSize = sizeof(RawDataType);
@@ -503,7 +503,9 @@ void Us4OEMImpl::validate(const std::vector<TxRxParametersSequence> &sequences, 
     auto nFirings = getNumberOfFirings(sequences);
     auto nTriggers = getNumberOfTriggers(sequences, rxBufferSize);
 
-    ARRUS_REQUIRES_AT_MOST(nFirings, 1024, format("Exceeded the maximum ({}) number of timeoutIds: {}", 1024, nFirings));
+    auto maxFirings = descriptor.getTxRxSequenceLimits().getMaxNumberOfFirings();
+
+    ARRUS_REQUIRES_AT_MOST(nFirings, maxFirings, format("Exceeded the maximum ({}) number of timeoutIds: {}", maxFirings, nFirings));
     const auto maxSequenceSize = descriptor.getTxRxSequenceLimits().getSize().end();
     ARRUS_REQUIRES_AT_MOST(nTriggers, maxSequenceSize,
                            format("Exceeded the maximum ({}) number of triggers: {}", maxSequenceSize, nTriggers));
@@ -612,8 +614,6 @@ uint32 Us4OEMImpl::getFirmwareVersion() { return ius4oem->GetFirmwareVersion(); 
 
 uint32 Us4OEMImpl::getTxFirmwareVersion() { return ius4oem->GetTxFirmwareVersion(); }
 
-uint32_t Us4OEMImpl::getTxOffset() { return ius4oem->GetTxOffset(); }
-
 uint32_t Us4OEMImpl::getOemVersion() { return ius4oem->GetOemVersion(); }
 
 void Us4OEMImpl::checkState() { this->checkFirmwareVersion(); }
@@ -628,7 +628,7 @@ void Us4OEMImpl::setTestPattern(RxTestPattern pattern) {
 
 std::pair<uint32_t, float> Us4OEMImpl::getTxStartSampleNumberAfeDemod(float ddcDecimationFactor) {
     //DDC RX offset (valid data offset)
-    uint32_t txOffset = ius4oem->GetTxOffset();
+    uint32_t txOffset = descriptor.getSampleTxStart();
     uint32_t rxOffset = 34u + (uint32_t)(16 * ddcDecimationFactor);
     uint32_t filterDelay = (uint32_t)(8 * ddcDecimationFactor);
 
@@ -768,14 +768,16 @@ size_t Us4OEMImpl::getNumberOfFirings(const std::vector<TxRxParametersSequence> 
 void Us4OEMImpl::setTxDelays(const std::vector<bool> &txAperture, const std::vector<float> &delays, uint16 firingId,
                              size_t delaysId, const std::unordered_set<ChannelIdx> &maskedChannelsTx) {
     ARRUS_REQUIRES_EQUAL_IAE(txAperture.size(), delays.size());
+    std::vector<float> delaysToBeApplied(txAperture.size());
     for (uint8 ch = 0; ch < ARRUS_SAFE_CAST(txAperture.size(), uint8); ++ch) {
         bool bit = txAperture.at(ch);
         float delay = 0.0f;
         if (bit && !setContains(maskedChannelsTx, static_cast<ChannelIdx>(ch))) {
             delay = delays.at(ch);
         }
-        ius4oem->SetTxDelay(ch, delay, firingId, delaysId);
+        delaysToBeApplied.at(ch) = delay;
     }
+    ius4oem->SetTxDelays(delaysToBeApplied, firingId, delaysId);
 }
 
 void Us4OEMImpl::clearCallbacks() {